xidmon.cpp: Ignore XIDeviceDisabled
At the moment, the primary use case of xireloadd is when all input
devices are hotplugged at once -- KVM switching and monitor poweroff/on
are the big ones. In that situation, it is fair for .xinputrc to expect
that all referenced input devices exist.

By listening to XIDeviceDisabled, we run .xinputrc when those devices
are removed. This will cause errors in .xinputrc if e.g. it calls
`xinput` on a particular device that has been disconnected. We don't
even care about this situation in the first place, so just stop
listening to XIDeviceDisabled.

Signed-off-by: Richard Tollerton <richard.tollerton@gmail.com>